+++ src/components/Button.jsx
@@ -1,7 +1,7 @@
-export default function Button({ type = 'button', text, className, clickHandler }) {
+export default function Button({ type = 'button', text, className, clickHandler, id }) {
   return (
-    <button className={className} type={type} onClick={clickHandler}>
-      <span className="btn_text">{text}</span>
+    <button className={className} type={type} onClick={clickHandler} data-id={id}>
+      <span className="btn-text">{text}</span>
     </button>
   );
 }